Add Sequencer logic
This adds a Sequencer class, which will do the job of the various sequencers in the NES APU such as the Frame Sequencer, or the Pulse Channel's duty sequencer.

The sequencer can switch between different sequence sets, and notify listeners who are listening to a sequence identified by its ID. A sequence with the same ID can exist in different sets.

This allows easily switching between the two modes (4-Step and 5-Step) for example in the Frame Sequencer