Change divider behavior to match hardware
Some sources model the divider to reload the same clock it hits 0. The hardware reloads on the tick afterwards. The latter approach is more compatible with the APU registers, as they set the reload value in the way intended for the hardware. Before, I had to add 1 everywhere, but then there would be an issue where the pulse channel's divider is set in two parts.