Update README: Audio is now supported 😏
Alright, folks, brace yourselves for a commit that will surely redefine the boundaries of audacity and humility in the realm of NES emulation. You might recall that our README previously had a line that gleefully declared, "Audio support? Nah, we don't do that here." Well, gather 'round because the tides of destiny have shifted, and we've decided to make a radical change – we've implemented audio support. Don't blink; you might just miss it!

You see, it took us (okay, let's be honest, it was just me, your friendly neighborhood coder) about as much time as it takes to microwave a frozen burrito to turn the tables and give our emulator the auditory makeover it so desperately didn't need. But hey, who's counting?

With this epic one-line alteration, we've taken the bold step of removing that disclaimer about audio absence from our README. It's gone, vanished, vamoosed, as if it never existed. Why, you ask? Because, dear user, the audio gods have graced us with their presence, and our emulator can now churn out sound like a malfunctioning kazoo caught in a windstorm. It's like Mozart playing on a Casio keyboard – in a hurricane!

Picture this: a lone programmer, armed with a handful of energy drinks and a suspiciously large collection of rubber duckies, decided to conquer the untamed wilderness of audio emulation. In a daring feat that left the world of NES emulation stunned, we accomplished this herculean task in approximately the time it takes to finish a cup of instant noodles. But hey, who's rushing?

So, in all seriousness (or not), we present to you our groundbreaking audio support. It may not rival the sound quality of a concert grand piano or a professional symphony orchestra, but it'll make you feel like you're playing NES games through a vintage radio submerged in a fish tank – and isn't that the dream?

We humbly (well, sort of) invite you to test out our latest masterpiece. Please note that your ears may experience a rollercoaster of emotions ranging from laughter to bewilderment, but we guarantee it's an experience you won't forget. And who knows, with enough time, maybe one day our audio will be as legendary as the NES itself. Until then, enjoy the quirky charm of our NES emulator, now with sound that'll make you question your life choices.