Simplify relative addressing calculation
At long last "//todo: this is dumb" has been taken care of. It was indeed pretty dumb to manually check for the 8th bit to determine whether to add or subtract. Instead I am now casting the value to a byte, which inherently becomes negative and maths correctly when it needs to