Restructure Project
This fixes a couple of things. For one, resource are no longer contained in a source folder, meaning they will no longer be copied to the output folder twice. Also, tests are now separated so they will not be compiled on a normal build, so the gigabyte of test data isn't included in builds either. I also removed unused images from resources and just shoved them into project root for now