Refactor clocking / timing
Emulating the master clock signal is a bit slow. This should be faster, and ties the whole timing to the CPU clock. The APU needed a lot of adjustment, as it was driven by a division on the master clock line. Now it is clocked every cpu cycle and can do its get/put cycle shenanigans